Fix v7.1.0 regression
This code works with 7.0.x and broken in 7.1.0
TypedError({
   type: 'myError',
   message: 'myError',
   length: null,
});

Reason: v7.0.x uses xtend package which not use 'use strict'.
